\name{mapInfo}
\alias{mapInfo}
\docType{data}
\title{
mafInfo - example data}
\description{
 a data frame containing the gene information for the variants in the study.
}
\usage{data("mapInfo")}
\format{
  A data frame of 20 observations on the following 2 variables.
  \describe{
    \item{\code{GENE}}{The gene name}
    \item{\code{SNP}}{An unique SNP identifier}
  }
}


\examples{
data(mapInfo)
}
\keyword{datasets}
