\name{GESE-internal}
\alias{GESE-internal functions}
\docType{package}
\title{
GESE package internal functions
}
\description{GESE package internal functions.
}
\details{
\code{computeP_resampling}
\code{findIntermediateFounder}
\code{findMostRecentCommonFounder}
\code{findMostRecentCommonFounderControl}
\code{getFounder}
\code{getProb}
\code{getPvalue_resampling}
\code{getTranProb_dv}
\code{isRelated}
\code{oneSetSim}
\code{segProb}
\code{getProb_weight}
}
\author{
Dandi Qiao

Maintainer: \packageMaintainer{GESE}
}
\references{
Qiao, D. Lange, C., Laird, N.M., Won, S.,  Hersh, C.P., et al. (2017). \emph{Gene-based segregation method for identifying rare variants for family-based sequencing studies.} Genet Epidemiol 41(4):309-319. DOI:10.1002/gepi.22037.
}

\keyword{ package }
\seealso{
\code{\link[GESE:GESE-package]{GESE}}
}
